+++ inst/shinyApp/ui/SidebarPanel.R
@@ -1,5 +1,5 @@
 # Sidebar panel ----------------------------------------------------------------
-spFunc <- function() {
+sidebar_panel <- function() {
     sidebarPanel(
         h3("Sidebar Panel")
     )